+++ android/quest/src/main/java/org/smartregister/fhircore/quest/data/QuestPatientRepository.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2021 Ona Systems, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.smartregister.fhircore.quest.data
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.google.android.fhir.FhirEngine
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+import org.hl7.fhir.r4.model.Patient
+import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
+import org.smartregister.fhircore.engine.util.DispatcherProvider
+
+class QuestPatientRepository(
+  private val fhirEngine: FhirEngine,
+  private val dispatcherProvider: DispatcherProvider = DefaultDispatcherProvider
+) {
+
+  fun fetchDemographics(patientId: String): LiveData<Patient> {
+    val data = MutableLiveData<Patient>()
+    CoroutineScope(dispatcherProvider.io()).launch {
+      data.postValue(fhirEngine.load(Patient::class.java, patientId))
+    }
+    return data
+  }
+}