+++ android/quest/src/test/java/org/smartregister/fhirecore/quest/data/QuestPatientRepositoryTest.kt
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2021 Ona Systems, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.smartregister.fhirecore.quest.data
+
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import com.google.android.fhir.FhirEngine
+import io.mockk.coEvery
+import io.mockk.mockk
+import org.hl7.fhir.r4.model.HumanName
+import org.hl7.fhir.r4.model.Patient
+import org.hl7.fhir.r4.model.StringType
+import org.junit.Assert
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.robolectric.annotation.Config
+import org.smartregister.fhircore.anc.coroutine.CoroutineTestRule
+import org.smartregister.fhircore.quest.data.QuestPatientRepository
+import org.smartregister.fhirecore.quest.robolectric.RobolectricTest
+import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
+
+@Config(shadows = [QuestApplicationShadow::class])
+class QuestPatientRepositoryTest : RobolectricTest() {
+
+  private lateinit var repository: QuestPatientRepository
+
+  @get:Rule var instantTaskExecutorRule = InstantTaskExecutorRule()
+  @get:Rule var coroutinesTestRule = CoroutineTestRule()
+
+  @Before
+  fun setUp() {
+    val fhirEngine = mockk<FhirEngine>()
+
+    coEvery { fhirEngine.load(Patient::class.java, "1") } returns
+      Patient().apply {
+        name =
+          listOf(
+            HumanName().apply {
+              given = listOf(StringType("john"))
+              family = "doe"
+            }
+          )
+      }
+
+    repository = QuestPatientRepository(fhirEngine, coroutinesTestRule.testDispatcherProvider)
+  }
+
+  @Test
+  fun testFetchDemographicsShouldReturnTestPatient() {
+    val patient = repository.fetchDemographics("1").value
+    Assert.assertEquals("john", patient?.name?.first()?.given?.first()?.value)
+    Assert.assertEquals("doe", patient?.name?.first()?.family)
+  }
+}